debayer stratery is running

+++ hilaris/hilaris/includes/Debayer.h
@@ -8,7 +8,7 @@ class Debayer
 {
 	public:
 		virtual enum EnOscPictureType getType() = 0;
-		virtual Image* getObject(uint8 width, uint8 height) = 0;
+		virtual Image* getObject(uint16 width, uint16 height) = 0;
 		virtual bool debayer(RawImage* raw, Image* image) = 0;
 	protected:
 		Debayer(){}
